Feat : Core Data Model 생성 및 MemoProvider 변경 - Memo Model type을 생성 기존 struct Memo 타입 제거 - MemoProvider 내로 CoreData Store Container 이동
