+++ CloudNotes/CloudNotes/Entry/SceneDelegate.swift
@@ -53,7 +53,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     // to restore the scene back to its current state.
     
     // Save changes in the application's managed object context when the application transitions to the background.
-    (UIApplication.shared.delegate as? AppDelegate)?.saveContext()
+    MemoProvider.shared.saveContext()
   }
