+++ CloudNotes/CloudNotes/Model/MemoProvider.swift
@@ -6,14 +6,43 @@
 //
 
 import UIKit
+import CoreData
 
 final class MemoProvider {
+  // MARK: - For Single Tone
+  static let shared: MemoProvider = MemoProvider()
+  private init() { }
+    
+  // MARK: - Core Data stack
+  lazy var persistentContainer: NSPersistentCloudKitContainer = {
+    let container = NSPersistentCloudKitContainer(name: "CloudNotes")
+    container.loadPersistentStores(completionHandler: { (storeDescription, error) in
+      if let error = error as NSError? {
+        fatalError("Unresolved error \(error), \(error.userInfo)")
+      }
+    })
+    return container
+  }()
+  
+  // MARK: - Core Data Saving support
+  func saveContext () {
+    let context = persistentContainer.viewContext
+    if context.hasChanges {
+      do {
+        try context.save()
+      } catch {
+        let nserror = error as NSError
+        fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
+      }
+    }
+  }
+  
+  // MARK: - Manage Memo Data
   private var memo: [Memo]?
+
   
   func getMockData() throws -> [Memo]? {
-    guard let dataAsset = NSDataAsset(name: "sample") else { return nil }
-    let data = try JSONDecoder().decode([Memo].self, from: dataAsset.data)
-    self.memo = data
-    return memo
+    
+    return nil
   }
 }
