+++ CloudNotes/CoreData/Memo+CoreDataClass.swift
@@ -1,0 +1,15 @@
+//
+//  Memo+CoreDataClass.swift
+//  CloudNotes
+//
+//  Created by 이영우 on 2021/06/17.
+//
+//
+
+import Foundation
+import CoreData
+
+@objc(Memo)
+public class Memo: NSManagedObject {
+
+}
